Final Project small amends
All these changes were lost to another version of the files, but now are
found.
Added description.
Added messages for the validation in spans.
Added social icons in the footer.
All external scripts and css-es are now accessed via CDNs.

In the script.js -
Added Cyrilic letters in the regex for names,
Added a # in  "#contact_submit_btn" : "#send_reservation";
Removed the code for google maps, due to the use of the iframe.